requirements.txt
streamlit==1.37.1
chardet==5.2.0
python-dotenv==1.0.1
